% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{genderapi-defunct}
\alias{genderapi-defunct}
\alias{get_gender_by_name}
\alias{get_gender_by_email}
\alias{get_gender_by_username}
\alias{get_gender_by_name_bulk}
\alias{get_gender_by_email_bulk}
\alias{get_gender_by_username_bulk}
\title{Defunct V1 functions}
\usage{
get_gender_by_name(...)

get_gender_by_email(...)

get_gender_by_username(...)

get_gender_by_name_bulk(...)

get_gender_by_email_bulk(...)

get_gender_by_username_bulk(...)
}
\arguments{
\item{...}{Ignored.}
}
\description{
The V1 functions of genderapi 1.x were removed in 2.0.0 because the V2
API uses different routes, request fields and responses. Calling one of
them raises an error that names the replacement:
}
\details{
\tabular{ll}{
   1.x \tab 2.0.0 \cr
   \code{get_gender_by_name()} \tab \code{\link[=genderapi_name]{genderapi_name()}} \cr
   \code{get_gender_by_email()} \tab \code{\link[=genderapi_email]{genderapi_email()}} \cr
   \code{get_gender_by_username()} \tab \code{\link[=genderapi_username]{genderapi_username()}} \cr
   \code{get_gender_by_name_bulk()}, \code{get_gender_by_email_bulk()}, \code{get_gender_by_username_bulk()} \tab \code{\link[=genderapi_batch]{genderapi_batch()}} \cr
}


The 1.x releases remain on CRAN's archive and on the \code{v1} branch of the
repository.
}
\keyword{internal}
