% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderapi.R
\name{get_gender_by_email}
\alias{get_gender_by_email}
\title{Get Gender by Email}
\usage{
get_gender_by_email(api_key, email, country = NULL, askToAI = FALSE)
}
\arguments{
\item{api_key}{Your GenderAPI.io API key.}

\item{email}{Email string to check.}

\item{country}{Optional two-letter country code (e.g. "US").}

\item{askToAI}{Logical. TRUE to force AI lookup, FALSE by default.}
}
\value{
A list with gender prediction result.
}
\description{
Get Gender by Email
}
