% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderapi.R
\name{get_gender_by_name}
\alias{get_gender_by_name}
\title{Get Gender by Name}
\usage{
get_gender_by_name(
  api_key,
  name,
  country = NULL,
  askToAI = FALSE,
  forceToGenderize = FALSE
)
}
\arguments{
\item{api_key}{Your GenderAPI.io API key.}

\item{name}{Name string to check.}

\item{country}{Optional two-letter country code (e.g. "US").}

\item{askToAI}{Logical. TRUE to force AI lookup, FALSE by default.}

\item{forceToGenderize}{Logical. TRUE to force analysis of nicknames.}
}
\value{
A list with gender prediction result.
}
\description{
Get Gender by Name
}
